requirements.txt
absl-py==0.6.1
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
cffi==1.11.5
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29
cytoolz==0.9.0.1
dask==1.0.0
decorator==4.3.0
dlib==19.16.0
easydict==1.9
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
html5lib==0.9999999
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
msgpack==0.5.6
networkx==2.2
numpy==1.15.4
opencv-python==3.4.3.18
pandas==0.23.4
pbr==5.1.1
Pillow==5.3.0
protobuf==3.6.1
pycparser==2.19
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
subprocess32==3.5.3
tb-nightly==1.13.0a20181125
tensorboardX==1.4
termcolor==1.1.0
tf-estimator-nightly==1.12.0.dev20181125
tf-nightly==1.13.0.dev20181125
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tqdm==4.28.1
Werkzeug==0.14.1
